build BardsStaking

+++ contracts/utils/Cobbs.sol
@@ -69,8 +69,8 @@ library LibCobbDouglas {
         n = LibFixedMath.exp(
             LibFixedMath.mulDiv(
                 LibFixedMath.ln(n),
-                int256(alphaNumerator),
-                int256(alphaDenominator)
+                int256(uint256(alphaNumerator)),
+                int256(uint256(alphaDenominator))
             )
         );
         // Compute
